requirements.txt
alembic==1.13.2
altair==5.4.0
asgiref==3.8.1
attrs==24.2.0
Authlib==1.3.1
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.1.1
azure-mgmt-subscription==3.1.1
babel==2.16.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
boto3==1.34.158
botocore==1.34.158
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
cryptography==42.0.8
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.1
django-allauth==0.63.6
django-crispy-forms==2.2
django-extensions==3.2.3
django-gsheets==0.0.10
django-heroku==0.3.1
django-import-export==4.1.1
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.2.0
eventlet==0.36.1
Flask==3.0.3
flask-babel==4.0.0
Flask-Compress==1.15
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
googlemaps==4.10.0
greenlet==3.0.3
gspread==6.1.2
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
jsonformatter==0.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
ldap3==2.9.1
libgravatar==1.0.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.4.0
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
mysqlclient==2.2.4
narwhals==1.3.0
numpy==2.0.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
paramiko==3.4.0
passlib==1.7.4
pillow==10.4.0
plotly==5.23.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-c==3.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
qrcode==7.4.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
SecretStorage==3.3.3
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
speaklater3==1.4
SQLAlchemy==2.0.32
sqlparse==0.5.0
sshtunnel==0.4.0
streamlit==1.37.1
tablib==3.5.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
watchdog==4.0.2
Werkzeug==3.0.3
whitenoise==6.7.0
wsproto==1.2.0
WTForms==3.1.2
zipp==3.19.2
zstandard==0.23.0
